Remove option to add found symbols to page context
Served no purpose and was impossible to get working problerly, as preprocessSource only has access to the component's file path, not the component.

This is a breaking change but nobody is using this plugin yet so we'll treat it as a hotfix and hope nobody notices. :) if you are reading this.